Refactoring done based on responsibility-driven: TaskManger: functionality of the application Menu: Navigation containing User input and System.out.println statements FileService: file-related (write, read) methods
